Gibbs sampling of variances and structure matrix (#12)
* Fix type annotation.

* Add sampling variances.

* Add tests for variance sampling shrinkage.

* Add tests for variance sampling shrinkage.

* Rename subpackage.

* Add structure sampling and a unit test for it.

* Fix type annotation.

* Add tests for sampling gamma.

* Fix comment.

+++ src/jnotype/logistic/logreg.py
@@ -60,7 +60,7 @@ def calculate_loglikelihood_matrix_from_variables(
     *,
     intercepts: Float[Array, " G"],
     coefficients: Float[Array, "G F"],
-    structure: Float[Array, "G F"],
+    structure: Int[Array, "G F"],
     covariates: Float[Array, "N F"],
     observed: Int[Array, "N G"],
 ) -> Float[Array, "N G"]:
